requirements.txt
# JustCodeWorks.EU - Clean Multi-Tenant PlatformDjango==5.2.7

# Core Django & Databasedjango-tenants==3.9.0

Django>=5.0.0psycopg2-binary==2.9.9

psycopg2-binary>=2.9.0python-decouple==3.8

Pillow==10.4.0

# Multi-tenancygunicorn==21.2.0

django-tenants>=3.6.1whitenoise==6.6.0

# Internationalization & Localization
django-rosetta>=0.9.9

# AI & OpenAI Integration
openai>=1.0.0
python-dotenv>=1.0.0

# Blog & Content Management
django-ckeditor>=6.7.0
django-taggit>=4.0.0
Pillow>=10.0.0

# Development & Production
django-debug-toolbar>=4.2.0
gunicorn>=21.2.0
whitenoise>=6.6.0

# API & Forms
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7

# Utilities
python-slugify>=8.0.0
django-extensions>=3.2.0